requirements.txt
aiogram
motor
dnspython
python-dotenv
aiohttp
googlemaps
Pillow
deep-translator
certifi
